+++ board/sw/Handler.hpp
@@ -16,22 +16,22 @@ struct Handler final
   Io::Proto::Get_persistent_config::Reply handle(Io::Proto::Get_persistent_config::Request const& req)
   {
     return {
-      .min_pos_ = static_cast<uint32_t>( round(ctx_.setpoints_.min_pos_ * 999) ),
-      .max_pos_ = static_cast<uint32_t>( round(ctx_.setpoints_.max_pos_ * 999) ),
-      .LED_brightness_ = static_cast<uint32_t>( round(ctx_.setpoints_.LED_brightness_ * 99) )
+      .min_pos_ = static_cast<uint32_t>( roundf(ctx_.setpoints_.min_pos_ * 999) ),
+      .max_pos_ = static_cast<uint32_t>( roundf(ctx_.setpoints_.max_pos_ * 999) ),
+      .LED_brightness_ = static_cast<uint32_t>( roundf(ctx_.setpoints_.LED_brightness_ * 99) )
     };
   }
 
   Io::Proto::Get_servo_position::Reply handle(Io::Proto::Get_servo_position::Request const& req)
   {
-    return { .pos_ = static_cast<uint32_t>( round(ctx_.last_reads_.position_ * 999) ) };
+    return { .pos_ = static_cast<uint32_t>( roundf(ctx_.last_reads_.position_ * 999) ) };
   }
 
   Io::Proto::Get_telemetry::Reply handle(Io::Proto::Get_telemetry::Request const& req)
   {
     return {
-      .eng_current_mA_ = static_cast<uint32_t>( round(ctx_.last_reads_.engine_current_A_ * 1000.0f ) ),
-      .vcc_voltage_mV_ = static_cast<uint32_t>( round(ctx_.last_reads_.vcc_V_            * 1000.0f ) ),
+      .eng_current_mA_ = static_cast<uint32_t>( roundf(ctx_.last_reads_.engine_current_A_ * 1000.0f ) ),
+      .vcc_voltage_mV_ = static_cast<uint32_t>( roundf(ctx_.last_reads_.vcc_V_            * 1000.0f ) ),
     };
   }
